pyproject.toml
[tool.poetry]
name = "lungtissues"
version = "0.1.0"
description = ""
authors = ["As17-01 <As17-01@yandex.ru>"]
readme = "README.md"
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.2"
pandas = "^2.1.4"
scipy = "^1.11.4"
loguru = "^0.7.2"
hydra-core = "^1.3.2"
hydra-slayer = "^0.4.1"
cloudpickle = "^3.0.0"
omegaconf = "^2.3.0"
pillow = "^10.1.0"
openslide-python = "^1.3.1"
scikit-image = "^0.22.0"
opencv-python = "^4.8.1.78"

[tool.poetry.dev-dependencies]
pytest-cov = "^3.0.0"
pytest = "^6.0"
black = "^22.3.0"
isort = "^5.8.0"
flake8 = "^3.9.2"
flake8-docstrings = "^1.6.0"
pep8-naming = "^0.12.1"
mypy = "^1.2.0"
codespell = "^2.0.0"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | blib2to3
  | tests/data
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.mypy]
ignore_missing_imports = true
strict_optional = true
files = "src/"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
